% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFAVAR_tvp.R
\name{bayesFAVAR_TVP}
\alias{bayesFAVAR_TVP}
\title{Estimates FAVAR-TVP model using MCMC sampler.}
\usage{
bayesFAVAR_TVP(Y, l = ncol(Y), n.factors, p = 1, nburn = 10000,
  nsim = 50000, tau = 40, beta.algorithm = "DK",
  reject.explosive = FALSE)
}
\arguments{
\item{Y}{Matrix or timeseries object}

\item{l}{Number of time series to vectorize; all ts to vectorize must come before all other series}

\item{n.factors}{Number of factors tu use for first l series}

\item{p}{Integer for the lag order (default is p=1)}

\item{nburn}{Number of MCMC draws used to initialize the sampler (defaults to 10000)}

\item{nsim}{Number of MCMC draws excluding burn-in (defaults to 50000)}

\item{tau}{Length of training sample used for determining prior parameters via OLS}

\item{beta.algorithm}{Algorithm for drawing time-varying VAR parameters. Either 'DK' for Durbin and Coopman 2002, or 'CC' for Carter and Cohn 1994. Defaults to 'DK'}

\item{reject.explosive}{Should the explosive (not stable) draws of VAR parameters be rejected? Defaults to 'FALSE'. For rationale consult Cogley and Sargent 2005. May significantly increase computtion time!}
}
\value{
\item{beta}{Draws of time-varying parameters (beta_t). 4D array [t x M x M*p+1 x draw]}
        \item{H}{Draws of observation equation error term covariance matrix H. 3D array [M x M x draw]}
        \item{Q}{Draws of state equation error term covariance matrix Q. 3D array [M x M x draw]}
        \item{L}{Draws of \Lambda. 3D array [n.factors x l x draw]}
        \item{F}{Draws of factors. 3D array }
        \item{R}{Draws of observation equation error term covariance matrix R.}
}
\references{
\itemize{
              \item \insertRef{KoopKorobilis2010}{bayesVAR}
              \item \insertRef{CC1994}{bayesVAR}
              \item \insertRef{DK2002}{bayesVAR}
              \item \insertRef{Bernanke2004}{bayesVAR}
            }
}

