% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR_tvp.R
\name{bayesVAR_TVP}
\alias{bayesVAR_TVP}
\title{Estimates Time-Varying Parameters VAR using MCMC sampler.}
\usage{
bayesVAR_TVP(Y, p = 1, nburn = 10000, nsim = 50000, tau = 40,
  beta.algorithm = "DK")
}
\arguments{
\item{Y}{Matrix or timeseries object}

\item{p}{Integer for the lag order (default is p=1)}

\item{nburn}{Number of MCMC draws used to initialize the sampler (defaults to 10000)}

\item{nsim}{Number of MCMC draws excluding burn-in (defaults to 50000)}

\item{tau}{Length of training sample used for determining prior parameters via OLS}

\item{beta.algorithm}{Algorithm for drawing time-varying VAR parameters. Either 'DK' for Durbin and Coopman 2002, or 'CC' for Carter and Cohn 1994. Defaults to 'DK'}
}
\value{
\item{beta}{Draws of time-varying parameters (beta_t). 4D array [t x M x Mp+1 x draw]} 
        \item{H}{Draws of observation equation error term covariance matrix H. 3D array [M x M x draw]} 
        \item{Q}{Draws of state equation error term covariance matrix Q. 3D array [M x M x draw]}
}
\description{
Estimates Time-Varying Parameters VAR using MCMC sampler.
}
\references{
\itemize{
              \item \insertRef{CC1994}{bayesVAR}
              \item \insertRef{DK2002}{bayesVAR}
            }
}

