% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR_tvp.R, R/bayesFAVAR_tvp.R, R/init.R
\name{impulse.response.bayesVAR_TVP}
\alias{impulse.response}
\alias{impulse.response.bayesFAVAR_TVP}
\alias{impulse.response.bayesVAR_TVP}
\title{Impulse response function for bayes VAR models}
\usage{
\method{impulse.response}{bayesVAR_TVP}(model, R = 20, t = model$t,
  orthogonal = TRUE, reorder = 1:model$n, plot = TRUE)

\method{impulse.response}{bayesFAVAR_TVP}(model, R = 20, t = model$t,
  orthogonal = TRUE, reorder = 1:model$n, plot = TRUE)

impulse.response(x, ...)
}
\description{
Impulse response function for bayes VAR models
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesVAR_TVP}: Impulse response for bayes VAR-TVP

\item \code{bayesFAVAR_TVP}: Impulse response for bayes FAVAR-TVP
}}

