% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR_tvp.R
\name{MinimumEntropy_weights}
\alias{MinimumEntropy_weights}
\title{Minimum entropy weights}
\usage{
MinimumEntropy_weights(pd, g, g.rhs)
}
\arguments{
\item{pd}{Simulations of predictive density [T+1 x n x N.sim]}

\item{g}{Transformation for pd used to express to new information. Must return vector of length p}

\item{g.rhs}{Vector of length p}
}
\value{
\item{pi.star}{Adjusted probabilities}
        \item{KLIC}{Kullback-Leibler information criterion}
        \item{gamma}{Vector of Lagrange multipliers}
        \item{optim.code}{An integer indicating why the optimization process terminated. See ?nlm for info}
}
\description{
Finds minimum entropy weights using Kullback-Leibler information criterion.
New prediction density conditions are expressed by function \eqn{g} and vector \eqn{g.rhs} such that
\deqn{\mathbb{E}g(pd) = g.rhs}{Eg(pd) = g.rhs}
}
\references{
\itemize{
              \item \insertRef{Robertson2005}{bayesVAR}
            }
}

