% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR_tvp.R, R/bayesFAVAR_tvp.R, R/init.R
\name{predictive.density.bayesVAR_TVP}
\alias{predictive.density}
\alias{predictive.density.bayesFAVAR_TVP}
\alias{predictive.density.bayesVAR_TVP}
\title{Calculates predictive density for given bayes MCMC VAR model and number of periods}
\usage{
\method{predictive.density}{bayesVAR_TVP}(model, T = 10,
  currentLevels = tail(model$y, 1))

\method{predictive.density}{bayesFAVAR_TVP}(model, T = 4,
  currentLevels = tail(model$y, model$p), n.sim = 10)

predictive.density(x, ...)
}
\arguments{
\item{model}{bayes VAR model}

\item{T}{number of time steps ahead to predict}
}
\description{
Calculates predictive density for given bayes MCMC VAR model and number of periods
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesVAR_TVP}: Predictive density for bayes TVP-VAR model.

\item \code{bayesFAVAR_TVP}: Predictive density for bayes TVP-VAR model.
}}

