% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR_tvp.R
\name{predictive.density}
\alias{predictive.density}
\alias{predictive.density.bayesVAR_TVP}
\title{Calculates predictive density for given bayes MCMC VAR model and number of periods}
\usage{
predictive.density(x, ...)

\method{predictive.density}{bayesVAR_TVP}(model, T = 10)
}
\arguments{
\item{model}{bayes VAR model}

\item{T}{number of time steps ahead to predict}
}
\description{
Calculates predictive density for given bayes MCMC VAR model and number of periods
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesVAR_TVP}: Predictive density for bayes TVP-VAR model.
}}

